modules/de/requirements.txt
apache-airflow-providers-amazon[s3]
apache-airflow-providers-postgres
pandas
sqlalchemy
boto3
psycopg2-binary 